refactor(util): move JAR loading to utils
Extracts exporter JAR loading utilities from the broker module to the
util module so that it can be reused in the gateway and other places to
load arbitrary external artifacts.
