+++ util/src/main/java/io/camunda/zeebe/util/jar/ExporterJarLoadException.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.broker.exporter.jar;
+package io.camunda.zeebe.util.jar;
 
 import java.io.IOException;
 import java.nio.file.Path;
