+++ util/src/main/java/io/camunda/zeebe/util/jar/ThreadContextUtil.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.broker.exporter.jar;
+package io.camunda.zeebe.util.jar;
 
 import io.camunda.zeebe.util.CheckedRunnable;
 import org.agrona.LangUtil;
